Xmake and a bunch of game stuff
XMAKE:
	- Adjustments to the compilation flags used for clang-cl / clang++
	- Asset stamp and shader file stamps now work correctly

ENGINE:
	- New macro "ENGINE_API" that defines the symbol to be exported
	- Engine API items have been modified to be ENGINE_API by default

GAME:
	- Camera has been added to game state
	- Better editor controls
	- Beginning the actual tile editor work

RHI:
	- Fixed a bug with the RHI_render_camera_t where the matrices were swapped (oops...)
	- Fixed the mouse delta code

VULKAN:
	- A bit redudant, but doubly set the slang compiler settings inside the shader creation code

+++ code/code_generator/athena/athena.cpp
@@ -207,11 +207,11 @@ DEBUG_print_lambda_data(AST_node_t *lambda_AST, u32 indent)
             {
                 case AST_EXPRESSION_VALUE_INT:
                 {
-                    printf("%ld", value.int_value);
+                    printf("%lld", (unsigned long long)value.int_value);
                 }break;
                 case AST_EXPRESSION_VALUE_UNSIGNED:
                 {
-                    printf("%lu", value.unsigned_value);
+                    printf("%llu", (unsigned long long)value.unsigned_value);
                 }break;
                 case AST_EXPRESSION_VALUE_FLOAT:
                 {
@@ -286,11 +286,11 @@ DEBUG_print_structure_members(AST_node_t *structure, u32 indent)
             {
                 case AST_EXPRESSION_VALUE_INT:
                 {
-                    printf("%ld", value.int_value);
+                    printf("%lld", (unsigned long long)value.int_value);
                 }break;
                 case AST_EXPRESSION_VALUE_UNSIGNED:
                 {
-                    printf("%lu", value.unsigned_value);
+                    printf("%llu", (unsigned long long)value.unsigned_value);
                 }break;
                 case AST_EXPRESSION_VALUE_FLOAT:
                 {
@@ -333,11 +333,11 @@ DEBUG_print_enum_data(AST_node_t *enum_AST, u32 indent)
             {
                 case AST_EXPRESSION_VALUE_INT:
                 {
-                    printf("%ld", value.int_value);
+                    printf("%lld", (unsigned long long)value.int_value);
                 } break;
                 case AST_EXPRESSION_VALUE_UNSIGNED:
                 {
-                    printf("%lu", value.unsigned_value);
+                    printf("%llu", (unsigned long long)value.unsigned_value);
                 } break;
                 case AST_EXPRESSION_VALUE_FLOAT:
                 {
@@ -1560,11 +1560,11 @@ output_type_info_member_data(string_builder_t *builder, code_type_t *type, AST_n
             {
                 case AST_EXPRESSION_VALUE_INT:
                 {
-                    c_string_builder_sprintf(builder, "\t\t\t\t.int64 = %ld,\n", current_member->expression.value.int_value);
+                    c_string_builder_sprintf(builder, "\t\t\t\t.int64 = %lld,\n", current_member->expression.value.int_value);
                 }break;
                 case AST_EXPRESSION_VALUE_UNSIGNED:
                 {
-                    c_string_builder_sprintf(builder, "\t\t\t\t.u64 = %lu,\n", current_member->expression.value.unsigned_value);
+                    c_string_builder_sprintf(builder, "\t\t\t\t.u64 = %llu,\n", current_member->expression.value.unsigned_value);
                 }break;
                 case AST_EXPRESSION_VALUE_FLOAT:
                 {
@@ -1641,11 +1641,11 @@ athena_handle_type_info(const char *char_filepath, const char *output_path, bool
                         {
                             case AST_EXPRESSION_VALUE_INT:
                             {
-                                printf("%ld", AST->expression.value.int_value);
+                                printf("%lld", (unsigned long long)AST->expression.value.int_value);
                             }break;
                             case AST_EXPRESSION_VALUE_UNSIGNED:
                             {
-                                printf("%lu", AST->expression.value.unsigned_value);
+                                printf("%llu", (unsigned long long)AST->expression.value.unsigned_value);
                             }break;
                             case AST_EXPRESSION_VALUE_FLOAT:
                             {
